[FIX] account_payment_pro_receiptbook: Fix Multiple creation from payment register
•  Moves the override to the correct wizard directory
•  Adds _get_receiptbook() method to find the appropriate receiptbook based on partner type and company
•  Generates proper payment names using the receiptbook's sequence and document type prefix (e.g., "REC 00001")
•  Falls back to "/" only when no receiptbook is found
