+++ account_payment_pro_receiptbook/models/__init__.py
@@ -6,4 +6,3 @@ from . import account_chart_template
 from . import res_company
 from . import res_config_setting
 from . import account_journal
-from . import account_payment_register
