+++ account_payment_pro_receiptbook/wizard/__init__.py
@@ -1,1 +1,2 @@
 from . import account_resequence
+from . import account_payment_register
